Eradicate TensorVector::operator[]
Part of the effort to unify batch data structures.

operator[] returns only proxy, having the Tensor
in its interface allows for metadata desync between
sample and the whole batch.

Rework the operator[] to return simple view that
obviously don't allow to change the allocation
(or type, shape etc) of the underlying sample.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>