Adjust configuration for autoformatting with black
Temporarily disable the linter in this commit, before the whole
repository is adjusted to new formatting.

Add black to third party deps listing to include it in the deps update
procedure.

Use latest version but require only the one for given year according to
the compabilty policy.

Remove obsolete configs related to previous linter or exceptions related
to the formatting of some parts of DALI.

Keep the separate config with exceptions for autograph to limit the code
changes.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>

Remove unused docstring limit, add W503 and W504

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>